Fix docker compose config
The existing config was not correctly forwarding requests to the web
server. Since the web server is intentionally meant to be very basic, it
doesn't have SSL support so the nginx server should not to try to force
SSL on incoming requests, otherwise they will fail.

Also removing the port binding for the web server itself as it doesn't
need to be accessible from the host directly.