Add Js to mask email addresses and copy to Docker
Because there are GDPR requirements to prevent sensitive personal
information from being sent to third party systems we want to ensure
that any email addresses that are sent as query params do not appear in
the nginx logs. Ngnix provides an implementation of JS that can be
integrated with the log system, to check and modify any log data. The
maskQueryParams.js file is being added to check the request query params
against a very complicated email regex and replace any values that match
with an EMAIL FILTERED notice.

JS module docs: https://www.nginx.com/blog/introduction-nginscript/