Optionally disable query param logging
Because we have GDPR requirements to be able to delete a users
personally identifiable information (PII) and some of that information
(like email addresses) could be passed into our systems as query
parameters, we want the ability to disable Nginx query parameter
logging. This is a lot easier than trying to identify all possible
instances of PII that come into our systems and filter them out
explicitly.