Remove http_referer from nginx logs
The http_referer was still logging the query params of the request that
triggered the additional requests (in testing this was going to the main
page triggering the request for the favicon). Since we don't use the
http_referer info for any purpose it's being removed to avoid logging
any sensitive data in the query params.